Fix bug in conditional for system_get_date_types().
If no `$type` argument is used, it makes sense that we would expect a
single format and would adjust return array by adding keys directly to
the base array itself. This looks reversed from the current logic which
is resulting in additional array keys with empty values like `type`,
`title` and `module` getting added to the root of the array when no
`$type` argument is used. This throws errors in some contrib modules
that use this function, like `workflow_node`.

Fixes #6556